Introduce *PolicyManager* for native app lifecycle
:Release Notes:
Introduce *PolicyManager* for native app lifecycle

:Detailed Notes:
Introduce *PolicyManager* for native app lifecycle

:Testing Performed:

:QA Notes:

:Issues Addressed:

Reviewed-on: http://gpro.lge.com/c/webosose/sam/+/269736
Tested-by: Sangwoo Kang <sangwoo82.kang@lge.com>

+++ src/util/Time.h
@@ -23,7 +23,7 @@ using namespace std;
 
 class Time {
 public:
-    static double getCurrentTime();
+    static long long getCurrentTime();
     static string generateUid();
 
     Time();
